+++ outcome/src/pages/index/properties/default.vue
@@ -1,5 +1,5 @@
 <template>
-    <layout :title="name">
+    <layout :title="name" :copy="copy">
         <q-text mode="compact">{{data | defaultString}}</q-text>
     </layout>
 </template>
@@ -9,6 +9,13 @@ import BaseExtends from "./base/base-extends.js";
 
 export default {
     extends: BaseExtends,
+    computed: {
+        copy(){
+            return [
+                { value: typeof this.data==="object"?JSON.stringify(this.data):this.data }
+            ]
+        }
+    },
     filters: {
         defaultString(value){
             if(typeof value === "object"){