+++ docs/CHANGELOG.md
@@ -3,7 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-<!-- ## [Unreleased] -->
+<!--
+## [Unreleased]
+-->
+
+## [1.1.0] - 2020-05-17
+
+### Added
+
+- Add Select All/Unselect All to artboards select page.
+- The Copy function now can be used on some properties.
+
+### Changed
+
+- Update the version of `@qiqi1996/qi-design-vue`.
 
 ## [1.0.2] - 2020-05-17